# openapi.yaml
openapi: 3.1.0
info:
  title: 山海灵感便利店领域 API
  version: 2.0.0
  description: App 用户接口与管理端事务动作。字段统一 snake_case，金额统一十进制字符串。
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /payments/{paymentNo}:
    get:
      tags: [Payments]
      security: [{ UserBearer: [] }]
      parameters: [{ name: paymentNo, in: path, required: true, schema: { type: string } }]
      summary: 查询本人支付单
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /payments/callback/{channel}:
    post:
      tags: [Payments]
      summary: 支付渠道 HMAC 回调
      parameters:
        - { name: channel, in: path, required: true, schema: { type: string, enum: [wechat, alipay, apple] } }
        - { name: x-signature, in: header, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [payment_no, trade_no, status], properties: { payment_no: { type: string }, trade_no: { type: string }, status: { type: string, enum: [paid, failed] }, paid_at: { type: string, format: date-time } } } } } }
      responses: { "200": { $ref: "#/components/responses/Success" }, "401": { $ref: "#/components/responses/Error" } }
...
  /refunds/callback/{channel}:
    post:
      tags: [Payments]
      summary: 退款渠道 HMAC 回调
      parameters:
        - { name: channel, in: path, required: true, schema: { type: string, enum: [wechat, alipay, apple] } }
        - { name: x-signature, in: header, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [refund_no, status], properties: { refund_no: { type: string }, status: { type: string, enum: [processed, failed] } } } } } }
      responses: { "200": { $ref: "#/components/responses/Success" } }
...
  /resources/{resource}:
    get:
      security: [{ AdminBearer: [] }]
      summary: 管理端通用资源分页
      parameters: [{ name: resource, in: path, required: true, schema: { type: string } }]
      responses: { "200": { $ref: "#/components/responses/List" } }
    post:
      security: [{ AdminBearer: [] }]
      summary: 管理端新增资源
      parameters: [{ name: resource, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
      responses: { "201": { $ref: "#/components/responses/Success" } }
components:
  securitySchemes:
    UserBearer: { type: http, scheme: bearer, description: 登录返回的 App 会话 Token }
    AdminBearer: { type: http, scheme: bearer, description: 环境变量 API_TOKEN }
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties: { code: { type: string }, message: { type: string }, details: {} }
        requestId: { type: string }
  responses:
    Success:
      description: 成功
      content: { application/json: { schema: { type: object, properties: { data: {} } } } }
    List:
      description: 分页成功
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { type: array, items: {} }
              meta: { type: object, properties: { page: { type: integer }, pageSize: { type: integer }, total: { type: integer } } }
    Error:
      description: 请求失败
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }